feat(ui): flip default turns padding to { 0, 1 }
The previous default of { left = 1, right = 0 } pushed the turn
indicator one column away from the sign column and left no gap before
the buffer text. Swapping to { left = 0, right = 1 } hugs the sign
column and gives the content room to breathe, which reads better in
practice.

+++ lua/flemma/config/schema.lua
@@ -86,8 +86,8 @@ return s.object({
     enabled = s.boolean(true),
     padding = s.union(
       s.object({
-        left = s.integer(1),
-        right = s.integer(0),
+        left = s.integer(0),
+        right = s.integer(1),
       }),
       s.integer()
     ):coerce(function(value, _ctx)
